Update tests locations.
* Move tests out of stackinternal_test.go in favor of testing
  externally from stack_test package.
* Restrict TestCallerInlinedPanic to Go 1.9 and newer since
  older versions of Go did not return stack frames for inlined
  functions.
